Allow RPCs to have Docs
This patch adds a simple "Docs" option to our RPC service definitions, which is then added as Elixir `@doc` attributes, if specified. This can then be read at run-time for self-documenting APIs, e.g.

Note: we do our best to handle multi-line strings, which are poorly supported in Protobufs. Also, we could probably get argument types from the protobufs definition, but we currently do not do that and leave it up to the doc writer to do so.

```proto
service AccountService {
    rpc account(AccountRequest) returns (AccountResponse) {
        option (google.api.http) = { get: "/account" };
        option (google.api.docs) = {
            description:
                "Get user account information\n"
                "\n"
                "This adds cool things.\n"
                ""
            };
    }

    rpc post_account(AccountRequest) returns (AccountResponse) {
        option (google.api.http) = { post: "/account" };
    }
}
```

will generate

```elixir
defmodule API.Presidio.AccountService do
  use Protobuf.Service

  @doc """
  Get user account information

  This adds cool things.
  """
  rpc :account, API.Presidio.AccountRequest, API.Presidio.AccountResponse, get: "/account"
  rpc :post_account, API.Presidio.AccountRequest, API.Presidio.AccountResponse, post: "/account"
end
```

which can be read in an app as:

```elixir
Code.fetch_docs(API.Presidio.AccountService)
{:docs_v1, 94, :elixir, "text/markdown", :none, %{},
 [
   {{:function, :rpcs, 0}, 97, ["rpcs()"],
    %{"en" => "Get user account information\n\nThis adds cool things.\n"}, %{}}
 ]}
```